// app/components/Types/Types.ts
export interface contact {
  contactNo: string;
  type: string;
}

export interface mail {
  mailId: string;
  type: string;
}

export interface doc {
  DocId: string;
  type: string;
}

export interface Qualify {
  Collage: string;
  Studies: string;
  Score: string;
}

export interface Teacher {
 Name: string;
 Code: string;
 Role: string;
}

export interface Add {
 Address1: string;
 Address2: string;
 Address3: string;
 Code: string;
}